@@ -1,32 +1,77 @@
 {
-  "name": "dailyhot_api",
-  "version": "1.0.3",
-  "description": "一个今日热榜",
-  "main": "index.js",
+  "name": "dailyhot-api",
+  "version": "2.0.7.1",
+  "description": "An Api on Today's Hot list",
+  "keywords": [
+    "API",
+    "RSS"
+  ],
+  "homepage": "https://github.com/Spring2022abcjk/DailyHotApi_proj#readme",
+  "bugs": {
+    "url": "https://github.com/imsyy/DailyHotApi/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/imsyy/DailyHotApi.git"
+  },
+  "license": "MIT",
+  "author": "imsyy",
+  "main": "dist/index.js",
+  "declaration": true,
+  "declarationMap": true,
+  "sourceMap": true,
+  "files": [
+    "LICENSE",
+    "README.md",
+    "dist/**/*",
+    "!dist/logs/**/*"
+  ],
   "scripts": {
     "format": "prettier --write .",
-    "lint": "eslint . --ext .js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix",
-    "start": "node index.js",
-    "dev": "npx nodemon index.js",
-    "prd": "pm2 start index.js",
-    "build": "node index.js"
+    "lint": "eslint .",
+    "dev": "cross-env NODE_ENV=development tsx watch --no-cache src/index.ts",
+    "dev:cache": "cross-env NODE_ENV=development tsx watch src/index.ts",
+    "build": "tsc --project tsconfig.json",
+    "start": "cross-env NODE_ENV=development node dist/index.js"
   },
-  "author": "imsyy",
-  "license": "MIT",
+  "type": "module",
   "dependencies": {
-    "axios": "^1.3.4",
-    "cheerio": "1.0.0-rc.12",
-    "dotenv": "^16.0.3",
-    "koa": "^2.14.1",
-    "koa-bodyparser": "^4.3.0",
-    "koa-router": "^12.0.0",
-    "koa-static": "^5.0.0",
-    "koa-views": "^8.0.0",
-    "koa2-cors": "^2.0.6",
+    "@hono/node-server": "^1.13.7",
+    "axios": "^1.7.8",
+    "chalk": "^5.3.0",
+    "cheerio": "^1.0.0",
+    "dayjs": "^1.11.13",
+    "dotenv": "^16.4.6",
+    "feed": "^4.2.2",
+    "flatted": "^3.3.2",
+    "hono": "^4.6.12",
+    "iconv-lite": "^0.6.3",
+    "ioredis": "^5.4.1",
+    "md5": "^2.3.0",
     "node-cache": "^5.1.2",
-    "nodemon": "^2.0.22",
-    "eslint": "^8.48.0",
-    "eslint-plugin-vue": "^9.17.0",
-    "prettier": "^3.0.2"
+    "node-fetch": "^3.3.2",
+    "rss-parser": "^3.13.0",
+    "user-agents": "^1.1.379",
+    "winston": "^3.17.0"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.16.0",
+    "@types/cheerio": "^0.22.35",
+    "@types/md5": "^2.3.5",
+    "@types/node": "^22.10.1",
+    "@types/user-agents": "^1.0.4",
+    "@types/xml2js": "^0.4.14",
+    "@typescript-eslint/eslint-plugin": "^8.17.0",
+    "@typescript-eslint/parser": "^8.17.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^9.16.0",
+    "globals": "^15.13.0",
+    "prettier": "^3.4.1",
+    "tsx": "^4.19.2",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.17.0"
+  },
+  "engines": {
+    "node": ">=20"
   }
 }
